requirements.txt
streamlit>=1.38.0
anthropic>=0.25.0
python-dotenv>=1.0.0
faster-whisper>=1.0.0
requests>=2.31.0
pdfplumber>=0.10.0
python-docx>=1.1.0
Pillow>=10.0.0
supabase>=2.0.0
gtts==2.5.3
reportlab>=3.6.0
bcrypt==4.1.2
email-validator==2.1.0
